backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic-settings
psycopg2-binary
gunicorn
